Removed the Subsplit concern and moved its code into Segment
